Add Storage class and update Storage test accordingly
