+++ src/main/java/seedu/rainyDay/RainyDay.java
@@ -4,11 +4,9 @@ import seedu.rainyDay.command.Command;
 import seedu.rainyDay.data.FinancialReport;
 import seedu.rainyDay.data.FinancialStatement;
 
-import java.io.FileInputStream;
-import java.io.FileOutputStream;
+
 import java.io.IOException;
-import java.io.ObjectInputStream;
-import java.io.ObjectOutputStream;
+
 import java.util.ArrayList;
 import java.util.Scanner;
 
@@ -19,8 +17,9 @@ public class RainyDay {
     public static FinancialReport financialReport = new FinancialReport(new ArrayList<>());
 
     public static void main(String[] args) {
+
         try {
-            financialReport = new FinancialReport(loadFromFile(filePath));
+            financialReport = new FinancialReport(Storage.loadFromFile(filePath));
         } catch (IOException | ClassNotFoundException e) {
             System.out.println("No valid save file detected. Starting with empty financial data.");
         }
@@ -80,34 +79,7 @@ public class RainyDay {
         System.out.println("Please refer to the help table!");
     }
 
-    public static void writeToFile(FinancialReport statements, String filePath) {
-        try {
-            FileOutputStream writeData = new FileOutputStream(filePath);
-            ObjectOutputStream writeStream = new ObjectOutputStream(writeData);
 
-            ArrayList<FinancialStatement> storeStatements = new ArrayList<>();
-            for (int i = 0; i < statements.getStatementCount(); i += 1) {
-                storeStatements.add(statements.getFinancialStatement(i));
-            }
-
-            writeStream.writeObject(storeStatements);
-            writeStream.flush();
-            writeStream.close();
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-    }
-
-    public static ArrayList<FinancialStatement> loadFromFile(String filePath)
-            throws IOException, ClassNotFoundException {
-        FileInputStream readData = new FileInputStream(filePath);
-        ObjectInputStream readStream = new ObjectInputStream(readData);
-        ArrayList<FinancialStatement> statements = (ArrayList<FinancialStatement>) readStream.readObject();
-        readStream.close();
-        readData.close();
-
-        return statements;
-    }
 
 
 }
