+++ src/test/java/seedu/rainyDay/RainyDayTest.java
@@ -1,19 +1,14 @@
 package seedu.rainyDay;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertThrows;
-import static seedu.rainyDay.RainyDay.loadFromFile;
 
 import seedu.rainyDay.command.Command;
 import seedu.rainyDay.data.FinancialReport;
 import seedu.rainyDay.data.FinancialStatement;
-import org.junit.jupiter.api.Assertions;
+
 import org.junit.jupiter.api.Test;
 
-import java.io.File;
-import java.io.FileInputStream;
-import java.io.IOException;
-import java.io.ObjectInputStream;
+
 import java.util.ArrayList;
 
 class RainyDayTest {
@@ -65,43 +60,6 @@ class RainyDayTest {
         }
     }
 
-    public void writeToFileTest_fileExists() {
 
-        ArrayList<FinancialStatement> statements = new ArrayList<>();
-        FinancialReport financialReport = new FinancialReport(statements);
-        String filePath = "rainyDay.txt";
-        RainyDay.writeToFile(financialReport, filePath);
-        Assertions.assertTrue(new File(filePath).exists());
-    }
-
-    @Test
-    public void writeToFileTest_contentMatch() throws IOException, ClassNotFoundException {
-        ArrayList<FinancialStatement> statements = new ArrayList<>();
-        FinancialReport financialReport = new FinancialReport(statements);
-        String filePath = "rainyDay.txt";
-        financialReport.addStatement(new FinancialStatement("noodles", "in", 5));
-        RainyDay.writeToFile(financialReport, filePath);
-
-        FileInputStream readData = new FileInputStream(filePath);
-        ObjectInputStream readStream = new ObjectInputStream(readData);
-        @SuppressWarnings("unchecked")
-        ArrayList<FinancialStatement> data = (ArrayList<FinancialStatement>) readStream.readObject();
-        FinancialReport financialReportData = new FinancialReport(data);
-        readStream.close();
-
-        assertEquals(Command.generateReport(financialReport), Command.generateReport(financialReportData));
-    }
-
-    @Test
-    public void loadFromFile_emptyFile_iOExceptionThrown() {
-        String emptyFilePath = "thisFileDoesNotExist.txt";
-        assertThrows(IOException.class, () -> loadFromFile(emptyFilePath));
-    }
-
-    @Test
-    public void loadFromFile_invalidFileType_classNotFoundExceptionThrown() {
-        String invalidTypeFilePath = "test files/thisFileTypeIsInvalid.txt";
-        assertThrows(IOException.class, () -> loadFromFile(invalidTypeFilePath));
-    }
 
 }
